Update README.md
This is the FASTAPI implementation with background tasks functionality to produce texts within the images. This can be used to extract text from single image and when provided with multiple images, then it will provide the link button on the home page. This portal also uses the concepts of routers and repositories that helps us to build large scale projects